WooF JDBC driver now supports 'recid' reserved keyword in WHERE clauses and as a reserved column name when querying result sets. This is the fastest way to access a FileMaker row (faster than the primary key). recid's are longs (since you can have more than 2 billion records in a FileMaker database), so the best way to access it from a ResultSet is getLong("recid")
FmMetaData: Implemented getColumnPrvileges() to detect whether fields are read-only or no access. It does this by creating a temporary record and then deleting it.

Summary fields are now marked as read-only when calling getColumnPrivileges()

+++ trunk/Fmp360_JDBC/src/com/prosc/fmpjdbc/FmRecord.java
@@ -37,12 +37,12 @@ public class FmRecord {
 	private static final Logger log = Logger.getLogger( FmRecord.class.getName() );
 	private static final TimeZone defaultTimeZone = TimeZone.getDefault();
 	
-	private Integer recordId;
+	private Long recordId;
 	private Integer modCount;
 	private FmFieldList fieldList;
 	private String[] rawValues;
 
-	public FmRecord(FmFieldList fieldList, Integer recordId, Integer modCount) {
+	public FmRecord(FmFieldList fieldList, Long recordId, Integer modCount) {
 		this.fieldList = fieldList;
 		this.rawValues = new String[fieldList.getFields().size()];
 		this.recordId = recordId;
@@ -50,7 +50,7 @@ public class FmRecord {
 	}
 
 
-	public Integer getRecordId() {
+	public Long getRecordId() {
 		return recordId;
 	}
